fix: Scroll into messaage on quote reply
