+++ src/modules/Channel/context/hooks/useScrollToMessage.ts
@@ -2,11 +2,13 @@ import type { BaseMessage } from '@sendbird/chat/message';
 import { useCallback } from 'react';
 
 import { Logger } from '../../../../index';
+import { scrollToRenderedMessage } from '../utils';
 
 interface DynamicParams {
   setInitialTimeStamp: React.Dispatch<React.SetStateAction<number>>;
   setAnimatedMessageId: React.Dispatch<React.SetStateAction<number>>;
   allMessages: BaseMessage[];
+  scrollRef: React.RefObject<HTMLDivElement>;
 }
 
 interface StaticParams {
@@ -17,6 +19,7 @@ function useScrollToMessage({
   setInitialTimeStamp,
   setAnimatedMessageId,
   allMessages,
+  scrollRef,
 }: DynamicParams,
 { logger }: StaticParams,
 ): (createdAt: number, messageId: number) => void {
@@ -30,6 +33,7 @@ function useScrollToMessage({
         if (isPresent) {
           logger.info('Channel: scroll to message - message is present');
           setAnimatedMessageId(messageId);
+          scrollToRenderedMessage(scrollRef, createdAt);
         } else {
           logger.info('Channel: scroll to message - fetching older messages');
           setInitialTimeStamp(null);
